enable gov module
