+++ sims.mk
@@ -17,7 +17,7 @@ $(BINDIR)/runsim:
 test-sim-nondeterminism:
 	@echo "Running non-determinism test..."
 	@go test -mod=readonly $(SIMAPP) -run TestAppStateDeterminism -Enabled=true \
-		-NumBlocks=100 -BlockSize=200 -Commit=true -Period=0 -v -timeout 24h
+		-NumBlocks=10 -BlockSize=20 -Commit=true -Period=0 -v -timeout 24h
 
 
 test-sim-fullappsimulation:
